+++ sdk/keyvault/azure-security-keyvault-secrets/src/test/java/com/azure/security/keyvault/secrets/SecretClientTestBase.java
@@ -249,6 +249,9 @@ public abstract class SecretClientTestBase extends TestBase {
     public abstract void recoverDeletedSecretNotFound(HttpClient httpClient, SecretServiceVersion serviceVersion);
 
     @Test
+    public abstract void recoverDeletedSecretNotFoundWithPollingDuration(HttpClient httpClient, SecretServiceVersion serviceVersion);
+
+    @Test
     public abstract void backupSecret(HttpClient httpClient, SecretServiceVersion serviceVersion);
 
     void backupSecretRunner(Consumer<KeyVaultSecret> testRunner) {