+++ sdk/spring/azure-spring-boot/src/main/java/com/azure/spring/aad/implementation/AzureOAuth2Configuration.java
@@ -22,7 +22,7 @@ public abstract class AzureOAuth2Configuration extends WebSecurityConfigurerAdap
 
     protected OAuth2AccessTokenResponseClient<OAuth2AuthorizationCodeGrantRequest> accessTokenResponseClient() {
         DefaultAuthorizationCodeTokenResponseClient result = new DefaultAuthorizationCodeTokenResponseClient();
-        result.setRequestEntityConverter(new AuthzCodeGrantRequestEntityConverter(repo.defaultClient()));
+        result.setRequestEntityConverter(new AuthzCodeGrantRequestEntityConverter(repo.getAzureClient()));
         return result;
     }
 }