+++ sdk/spring/azure-spring-boot/src/test/java/com/azure/spring/autoconfigure/aad/AADOAuth2ConfigTest.java
@@ -15,19 +15,13 @@ import org.springframework.core.io.ClassPathResource;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.support.ResourcePropertySource;
 import org.springframework.mock.env.MockPropertySource;
-import org.springframework.security.oauth2.client.registration.ClientRegistration;
-import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserService;
 import org.springframework.test.context.support.TestPropertySourceUtils;
 import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
 
-import java.util.Arrays;
-import java.util.HashSet;
 import java.util.Map;
-import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class AADOAuth2ConfigTest {
     private static final String AAD_OAUTH2_MINIMUM_PROPS = "aad-backend-oauth2-minimum.properties";
@@ -119,40 +113,6 @@ public class AADOAuth2ConfigTest {
 
     }
 
-    @Test
-    public void testScopePropertyConfiguredWithDynamicPermissions() {
-        testContext = initTestContext("azure.activedirectory.scope=email");
-
-        final Environment environment = testContext.getEnvironment();
-        assertThat(environment.getProperty("azure.activedirectory.scope"))
-            .isEqualTo("email");
-
-        final ClientRegistrationRepository clientRegistrationRepository =
-            testContext.getBean(ClientRegistrationRepository.class);
-        final ClientRegistration clientRegistration = clientRegistrationRepository.findByRegistrationId("azure");
-        final Set<String> createdScopes = clientRegistration.getScopes();
-        final Set<String> expectedScopes = new HashSet<>(Arrays.asList("email", "openid", "profile",
-            "https://graph.microsoft.com/user.read"));
-        assertTrue(createdScopes.equals(expectedScopes));
-
-    }
-
-    @Test
-    public void testScopePropertyConfiguredWithStaticPermissions() {
-        testContext = initTestContext("azure.activedirectory.scope=1111/.default");
-
-        final Environment environment = testContext.getEnvironment();
-        assertThat(environment.getProperty("azure.activedirectory.scope"))
-            .isEqualTo("1111/.default");
-
-        final ClientRegistrationRepository clientRegistrationRepository =
-            testContext.getBean(ClientRegistrationRepository.class);
-        final ClientRegistration clientRegistration = clientRegistrationRepository.findByRegistrationId("azure");
-        final Set<String> createdScopes = clientRegistration.getScopes();
-        final Set<String> expectedScopes = new HashSet<>(Arrays.asList("1111/.default"));
-        assertTrue(createdScopes.equals(expectedScopes));
-
-    }
 
     private AnnotationConfigWebApplicationContext initTestContext(String... environment) {
         final AnnotationConfigWebApplicationContext context = new AnnotationConfigWebApplicationContext();