+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release History
 
 ## 12.10.0-beta.1 (Unreleased)
-
+- Added ability to get container client from blob clients and service client from container clients
 
 ## 12.9.0 (2020-11-11)
 - Fixed a bug where interspersed element types returned by page listing would deserialize incorrectly.